Revert "Update other event"
This reverts commit 9d9a97e25aa2148b82a9e1548801fa7a3d8ef382.
