+++ VideoApi/VideoApi.Events/Handlers/JoinedEventHandler.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using Microsoft.Extensions.Logging;
 using System.Threading.Tasks;
 using VideoApi.DAL.Commands;
@@ -21,27 +18,16 @@ namespace VideoApi.Events.Handlers
 
         public override EventType EventType => EventType.Joined;
 
-        protected override async Task PublishStatusAsync(CallbackEvent callbackEvent)
+        protected override Task PublishStatusAsync(CallbackEvent callbackEvent)
         {
             var participantState =  ParticipantState.Available;
             var room = RoomType.WaitingRoom;
-            var participantIds = new List<Guid>
-            {
-                SourceParticipant.Id
-            };
-            participantIds.AddRange(SourceParticipant.LinkedParticipants
-                .Where(p => p.Linked.State == ParticipantState.Available)
-                .Select(linkedParticipant => linkedParticipant.LinkedId));
+            var command = new UpdateParticipantStatusAndRoomCommand(SourceConference.Id, SourceParticipant.Id,
+                participantState, room, null);
             
-            foreach (var participantId in participantIds)
-            {
-                var command = new UpdateParticipantStatusAndRoomCommand(SourceConference.Id, participantId,
-                    participantState, room, null);
-            
-                _logger.LogInformation("Joined callback - {ConferenceId}/{ParticipantId}",
-                    SourceConference.Id, participantId);
-                await CommandHandler.Handle(command);
-            }
+            _logger.LogInformation("Joined callback - {ConferenceId}/{ParticipantId}",
+                SourceConference.Id, SourceParticipant.Id);
+            return CommandHandler.Handle(command);
         }
     }
 }
