+++ VideoApi/VideoApi.Events/Handlers/RoomParticipantJoinedEventHandler.cs
@@ -1,3 +1,6 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using VideoApi.DAL.Commands;
@@ -22,16 +25,29 @@ namespace VideoApi.Events.Handlers
         {
             var participantState =  ParticipantState.Available;
             var room = RoomType.WaitingRoom;
-            var updateParticipantCommand = new UpdateParticipantStatusAndRoomCommand(SourceConference.Id, SourceParticipant.Id,
-                participantState, room, null);
-            await CommandHandler.Handle(updateParticipantCommand);
-            var addParticipantToRoomCommand =
-                new AddParticipantToParticipantRoomCommand(SourceParticipantRoom.Id, SourceParticipant.Id);
-            
-            _logger.LogInformation("Room Participant Joined callback received - {ConferenceId}/{ParticipantId} - {ParticipantState} - {Room} {RoomLabel} - {SourceRoom}",
-                SourceConference.Id, SourceParticipant.Id, participantState, room, null, SourceParticipantRoom.Id);
 
-            await CommandHandler.Handle(addParticipantToRoomCommand);
+            var participantIds = new List<Guid>
+            {
+                SourceParticipant.Id
+            };
+
+            participantIds.AddRange(SourceParticipant.LinkedParticipants
+                .Where(p => p.Linked.State == ParticipantState.Available)
+                .Select(linkedParticipant => linkedParticipant.LinkedId));
+
+            foreach (var participantId in participantIds)
+            {
+                var updateParticipantCommand = new UpdateParticipantStatusAndRoomCommand(SourceConference.Id, participantId,
+                    participantState, room, null);
+                await CommandHandler.Handle(updateParticipantCommand);
+                var addParticipantToRoomCommand =
+                    new AddParticipantToParticipantRoomCommand(SourceParticipantRoom.Id, participantId);
+            
+                _logger.LogInformation("Room Participant Joined callback received - {ConferenceId}/{ParticipantId} - {ParticipantState} - {Room} {RoomLabel} - {SourceRoom}",
+                    SourceConference.Id, participantId, participantState, room, null, SourceParticipantRoom.Id);
+
+                await CommandHandler.Handle(addParticipantToRoomCommand);
+            }
         }
     }
 }
